feat: introduce GroupManager module
